+++ android/src/main/kotlin/com/dra11y/flutter/accessible_text_view/Extensions.kt
@@ -1,7 +1,9 @@
 package com.dra11y.flutter.accessible_text_view
 
+import android.graphics.Color
 import android.graphics.Typeface
 import android.os.Build
+import android.provider.CalendarContract.Colors
 import android.text.Spannable
 import android.text.SpannableString
 import android.text.TextPaint
@@ -22,14 +24,29 @@ fun Spannable.autoLinkify() {
     }
 }
 
-fun Spannable.formatSpans(typeface: Typeface) {
+fun Spannable.formatSpans(linkStyle: AccessibleTextViewOptions.TextStyle?) {
+    if (linkStyle == null) return
+
     val spans = getSpans(0, length, URLSpan::class.java)
         .sortedBy { getSpanStart(it) }
+    val typeface = linkStyle.resolveTypeface()
+    val color = linkStyle.color?.toColor() ?: Color.BLUE
+    val underlineColor = linkStyle.decorationColor?.toColor() ?: Color.BLUE
+
     for (span in spans) {
         val urlSpan = object : URLSpan(span.url) {
             override fun updateDrawState(ds: TextPaint) {
                 super.updateDrawState(ds)
                 ds.typeface = typeface
+                ds.color = color
+                ds.isUnderlineText = linkStyle.isUnderline
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                    ds.underlineColor = underlineColor
+                    linkStyle.decorationThickness?.let { thickness ->
+                        ds.underlineThickness = thickness
+                    }
+                    ds.isStrikeThruText = linkStyle.isLineThrough
+                }
             }
         }
         setSpan(