+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:accessible_text_view/accessible_text_view.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:native_segments/native_segments.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 void main() => runApp(const MaterialApp(home: TextViewExample()));
@@ -15,48 +16,62 @@ class TextViewExample extends StatefulWidget {
 class _TextViewExampleState extends State<TextViewExample> {
   bool isFlutterTextView = false;
 
-  Widget get accessibleTextView => AccessibleTextView(
-        html: """
-    <p>This is an AccessibleTextView widget. Its behavior is more correct on Android and more flexible on iOS.</p>
+  Widget get accessibleTextView => Padding(
+        padding: const EdgeInsets.only(top: 10),
+        child: AccessibleTextView(
+          html: html,
+          textStyle: normalStyle,
+          linkStyle: linkStyle,
+          brightness: Brightness.dark,
+          useFlutterTextScale: true,
+        ),
+      );
+
+  final String html1 = """
+  This is short text.
+""";
+
+  final String html = """
+    <p>This is an AccessibleTextView.</p>
     <p>This is a second paragraph with one <a href="https://google.com">Google</a> link.</p>
     <p>This is an example text with phone number 555-234-2345 with an e-mail link of test@example.com and a web link to <a href="https://google.com">Google</a>.
     On Android, TalkBack should indicate each link with a blip sound.
-    On iOS, VoiceOver should navigate each non-link and link chunk separately.</p>
+    On iOS, VoiceOver should navigate each paragraph separately, and provide a rotor and long-press popup menu (in VoiceOver only) for links.</p>
     <p>This component really should not be used for long text, because it is not designed for long text. For that, use a web view.</p>
-""",
-        // textColor: Theme.of(context).colorScheme.onSurface,
-        textColor: Colors.white,
-        linkWeight: FontWeight.bold,
-        linkColor: Colors.blueAccent,
-        appearance: AccessibleTextViewAppearance.dark,
-        fontSize: 14,
-      );
+  """;
 
-  final TextStyle flutterLinkStyle = const TextStyle(
+  final TextStyle normalStyle = const TextStyle(
+    color: Colors.white,
+    fontSize: 14,
+    height: 1.5,
+  );
+
+  final TextStyle linkStyle = const TextStyle(
     color: Colors.blueAccent,
     fontWeight: FontWeight.bold,
     decoration: TextDecoration.underline,
   );
 
-  final TextStyle normalStyle = const TextStyle(
-    color: Colors.white,
-    fontSize: 14,
-  );
-
   late final Widget flutterTextView = Text.rich(
     TextSpan(
-      text:
-          'This is a Flutter Text.rich. Its accessibility behavior with links is correct on iOS, but not Android.',
+      text: 'This is a Flutter Text.rich.',
       children: [
-        const TextSpan(
-            text:
-                '\n\nOn Android, the links are separate nodes, but are called out as buttons instead of links. Android users expect the text to be continuous, with a TalkBack menu of links.'),
+        const TextSpan(text: '\n\nThis is a second paragraph with one '),
+        TextSpan(
+          text: 'Google',
+          style: linkStyle,
+          recognizer: TapGestureRecognizer()
+            ..onTap = () async {
+              await launchUrlString('https://google.com');
+            },
+        ),
+        const TextSpan(text: ' link.'),
         TextSpan(
           text: '\n\nThis is an example text with phone number ',
           children: [
             TextSpan(
               text: '555-234-2345',
-              style: flutterLinkStyle,
+              style: linkStyle,
               recognizer: TapGestureRecognizer()
                 ..onTap = () async {
                   await launchUrlString('tel:555-234-2345');
@@ -65,7 +80,7 @@ class _TextViewExampleState extends State<TextViewExample> {
             const TextSpan(text: ' with an e-mail link of '),
             TextSpan(
               text: 'test@example.com',
-              style: flutterLinkStyle,
+              style: linkStyle,
               recognizer: TapGestureRecognizer()
                 ..onTap = () async {
                   await launchUrlString('mailto:test@example.com');
@@ -74,7 +89,7 @@ class _TextViewExampleState extends State<TextViewExample> {
             const TextSpan(text: ' and an external link to '),
             TextSpan(
               text: 'Google',
-              style: flutterLinkStyle,
+              style: linkStyle,
               recognizer: TapGestureRecognizer()
                 ..onTap = () async {
                   await launchUrlString('https://google.com');
@@ -109,14 +124,22 @@ class _TextViewExampleState extends State<TextViewExample> {
         backgroundColor: Colors.black38,
         body: Column(
           children: [
-            TextButton(
-              onPressed: () {
-                setState(() {
-                  isFlutterTextView = !isFlutterTextView;
-                });
-              },
-              child: Text(
-                  'Switch to ${isFlutterTextView ? 'Native AccessibleTextView' : 'Flutter Text.rich'}'),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: NativeSegments(
+                segments: [
+                  NativeSegment(title: 'AccessibleTextView'),
+                  NativeSegment(title: 'Flutter Text.rich'),
+                ],
+                style: const NativeSegmentsStyle(
+                  isDarkTheme: true,
+                ),
+                onValueChanged: (value) {
+                  setState(() {
+                    isFlutterTextView = value == 1;
+                  });
+                },
+              ),
             ),
             Expanded(
               flex: 3,